Integrated Auth0 into UserSettings and Zustand
